moving contract update to main NextDayController loop

+++ src/main/java/openwrestling/model/gameObjects/Contract.java
@@ -18,7 +18,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class Contract extends GameObject implements Serializable, iContract {
 
-    private int contractID;
+    private long contractID;
     private Promotion promotion;
     private Worker worker;
 
@@ -58,7 +58,7 @@ public class Contract extends GameObject implements Serializable, iContract {
     }
 
     public LocalDate getLastShowDate() {
-        if(lastShowDate == null) {
+        if (lastShowDate == null) {
             return startDate;
         }
         return lastShowDate;
